Refactor: adds Home context
This commit adds a Home page context that shares the state of the
generated id between the Form and the Card component, this made it
so that the Card component can be placed outside of the form
component making it more reusable and independent.

This commit also adds tests for the Introduction component as that
it has been fully refactored and finalized. More tests will be done
to the other components once they are refactored as well.

 Changes to be committed:
	deleted:    __tests__/404.test.tsx
	deleted:    __tests__/__snapshots__/404.test.tsx.snap
	new file:   __tests__/__snapshots__/index.spec.tsx.snap
	deleted:    __tests__/__snapshots__/index.test.tsx.snap
	renamed:    __tests__/index.test.tsx -> __tests__/index.spec.tsx
	modified:   src/components/Home/Card/Card.component.tsx
	modified:   src/components/Home/Form/Form.component.tsx
	new file:   src/components/Home/Home.context.tsx
	modified:   src/components/Home/Home.tsx
	modified:   src/components/Home/Introduction/Introduction.component.tsx
	new file:   src/components/Home/Introduction/Introduction.spec.tsx
	new file:   src/components/Home/Introduction/__snapshots__/Introduction.spec.tsx.snap
	modified:   src/components/shared/Footer.tsx
	modified:   src/pages/index.tsx